+++ test/client/unit/request/graphql/request.js
@@ -1538,6 +1538,9 @@ describe('GraphQL', () => {
       expect(graphQLRequest.adaptResponseBody(makeResponse('JCB')).creditCards[0].details.cardType).toBe('JCB');
       expect(graphQLRequest.adaptResponseBody(makeResponse('UNION_PAY')).creditCards[0].details.cardType).toBe('Union Pay');
       expect(graphQLRequest.adaptResponseBody(makeResponse('VISA')).creditCards[0].details.cardType).toBe('Visa');
+      expect(graphQLRequest.adaptResponseBody(makeResponse('ELO')).creditCards[0].details.cardType).toBe('Elo');
+      expect(graphQLRequest.adaptResponseBody(makeResponse('HIPER')).creditCards[0].details.cardType).toBe('Hiper');
+      expect(graphQLRequest.adaptResponseBody(makeResponse('HIPERCARD')).creditCards[0].details.cardType).toBe('Hipercard');
       expect(graphQLRequest.adaptResponseBody(makeResponse('SOLO')).creditCards[0].details.cardType).toBe('Unknown');
       expect(graphQLRequest.adaptResponseBody(makeResponse('UNKNOWN')).creditCards[0].details.cardType).toBe('Unknown');
       expect(graphQLRequest.adaptResponseBody(makeResponse()).creditCards[0].details.cardType).toBe('Unknown');